some code cleanup, reorganization and fixes
